+++ lua/haskell-tools/deps.lua
@@ -6,8 +6,10 @@ local M = {}
 -- @return the return value of on_available or on_not_available
 function M.if_available(modname, on_available, on_not_available)
   local has_mod, mod = pcall(require, modname)
-  if has_mod then
+  if has_mod and type(on_available) == 'function' then
     return on_available(mod)
+  elseif has_mod then
+    return on_available
   end
   if not on_not_available then
     return nil
@@ -18,14 +20,33 @@ function M.if_available(modname, on_available, on_not_available)
   return on_not_available
 end
 
-local function require_or_err(modname, err_msg)
+--@return unknown
+function M.require_or_err(modname, plugin_name)
   return M.if_available(
     modname,
     function(mod) return mod end,
-    function() error('haskell-tools: ' .. err_msg) end
+    function() error('haskell-tools: This plugin requires the ' .. plugin_name .. ' plugin.') end
   )
 end
 
-M.lspconfig = require_or_err('lspconfig', 'This plugin requires the neovim/nvim-lspconfig plugin')
+--@return boolean
+function M.has(modname)
+  return M.if_available(modname, true, false)
+end
+
+--@return boolean
+function M.has_telescope()
+  return M.has('telescope')
+end
+
+--@return unknown
+function M.require_telescope(modname)
+  return M.require_or_err(modname, 'nvim-telescope/telescope.nvim')
+end
+
+--@return unknown
+function M.require_plenary(modname)
+  return M.require_or_err(modname, 'nvim-lua/plenary.nvim')
+end
 
 return M